Fixed parsing problem when large activity blocks occurred in the JSON towards the end of a chunk. The earlier workaround involved adding a larger part of the following chunk to the leftover chunk of the previous loop, but that's risky since we don't know how long these blocks will potentially get.
Instead, mark keys that are parsed correctly, but are not interesting, as correctly parsed and not needing another attempt in the next run. This also decreases the need for a long overlap between consecutive chunks: we'll always be quite close to the end of one chunk and the extra required info is bound to be close to the start of the following chunk.