Didn't like the search API.
Changed it to take the target as an argument to find() instead of it being implicitly encoded in the comparison function. This makes it possible to use the same sarch algorithm with different targets to find instead of instantiating new ones, and is more consistent with what you'd expect the behaviour to be.